Fix references to service functions
