+++ breadbox/breadbox/compute/analysis_tasks.py
@@ -23,8 +23,11 @@ from breadbox.models.dataset import (
     ValueType,
 )
 from breadbox.schemas.custom_http_exception import ResourceNotFoundError, UserError
-
 from breadbox.schemas.dataset import MatrixDatasetIn
+from breadbox.service.labels import (
+    get_dataset_feature_labels_by_id,
+    get_dataset_feature_by_label,
+)
 
 from ..crud.types import get_dimension_type
 from ..crud import dataset as dataset_crud
@@ -170,9 +173,7 @@ def get_features_info_and_dataset(
     result_features: List[Feature] = []
     dataset_feature_ids: List[str] = []
     datasets: List[Dataset] = []
-    feature_labels_by_id = dataset_crud.get_dataset_feature_labels_by_id(
-        db, user, dataset
-    )
+    feature_labels_by_id = get_dataset_feature_labels_by_id(db, user, dataset)
     feature_indices = []
 
     for dataset_feat in dataset_features:
@@ -493,12 +494,12 @@ def create_cell_line_group(
 
         # Return the feature ID associated with the new dataset feature
         if use_feature_ids:
-            feature: DatasetFeature = dataset_crud.get_dataset_feature_by_label(
+            feature: DatasetFeature = get_dataset_feature_by_label(
                 db=db, dataset_id=dataset_id, feature_label=feature_label
             )
             return _format_breadbox_shim_slice_id(feature.dataset_id, feature.given_id)
         else:
-            dataset_feature = dataset_crud.get_dataset_feature_by_label(
+            dataset_feature = get_dataset_feature_by_label(
                 db, dataset_id, feature_label
             )
             return str(dataset_feature.id)
